// src/app/modules/userModule/user.interface.ts
import { Document, Types } from 'mongoose';

enum Gender {
    Male = 'male',
    Female = 'female',
    Other = 'other',
}

interface IUser extends Document {
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    password: string;
    status: string;
    isEmailVerified: boolean;
    verification?: {
        code: string;
        expireDate: Date;
    };
    role: string;
    isProvider:"googleAuth" | "appaleAuth";
    isSocial?: boolean;
    fcmToken?: string;
    profile: Types.ObjectId;

    // method declarations
    comparePassword(userPlanePassword: string): boolean;
    compareVerificationCode(userPlaneCode: string): boolean;
}

export default IUser;